Document IMU + friends
